+++ src/renderer/App.tsx
@@ -1,14 +1,14 @@
-import React from 'react';
-import './App.global.css';
-import Canvas from './components/Canvas';
-import { observer } from 'mobx-react';
+import React, { useEffect } from "react";
+import "./App.global.css";
+import MainLayout from "./containers/MainLayout";
+import { keyboardShortcutsReceiver } from "./events/receiveKeyboardShortcuts";
 
 function App() {
-	return (
-		<div className="CanvasContainer">
-			<Canvas />
-		</div>
-  );
+	useEffect(() => {
+		keyboardShortcutsReceiver(window);
+	}, []);
+
+	return <MainLayout />;
 }
 
-export default observer(App);
+export default App;