+++ src/renderer/components/Canvas/Cursor.tsx
@@ -1,23 +1,25 @@
-import React from "react"
-import { observer } from "mobx-react"
-import PaperStore from "renderer/stores/PaperStore"
+import React from "react";
+import { observer } from "mobx-react";
+import store from "renderer/stores/RootStore";
 
-const Cursor = () => (
+const Cursor = (props) => (
 	<div
-	className="cursor"
-	style={{
-		boxShadow: `inset 0 0 0 1px red`,
-		zIndex: 1000,
-		pointerEvents: "none",
-		width: PaperStore.cellSize.width,
-		height: PaperStore.cellSize.height,
-		transform: `translate(${PaperStore.positionInPixels.x}px, ${PaperStore.positionInPixels.y}px)`,
-		position: "absolute",
-		left: 0,
-		right: 0,
+		className="cursor"
+		style={{
+			boxShadow: `inset 0 0 0 1px red`,
+			zIndex: 1000,
+			pointerEvents: "none",
+			width: store.openProjects[props.id].cellSize.width,
+			height: store.openProjects[props.id].cellSize.height,
+			transform: `translate(${
+				store.openProjects[props.id].positionInPixels.x
+			}px,
+			${store.openProjects[props.id].positionInPixels.y}px)`,
+			position: "absolute",
+			left: 0,
+			right: 0,
 		}}
-	>
-	</div>
-)
+	></div>
+);
 
-export default observer(Cursor)
+export default observer(Cursor);