+++ src/renderer/stores/SettingsStore.ts
@@ -1,0 +1,13 @@
+import { makeAutoObservable } from "mobx";
+
+class SettingsStore {
+	rootStore: ClassDecorator;
+
+	constructor(rootStore) {
+		makeAutoObservable(this);
+		console.log("Settings");
+		this.rootStore = rootStore;
+	}
+}
+
+export default SettingsStore;