+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.0.2 (2021-10-21)
+
+- **FEATURES:** added command pattern functionality which saves any change to the canvas as a "command" that can be undoed and redoed.
+- **FEATURES:** Added Insert, Rotate and Flip commands.
+- **FEATURES:** Added FlexLayout as the main layout library
+- **FEATURES:** Added option to have multiple drawings open at the same time
+- **GENERAL:** Created rootstore which connects all other stores and keeps track of open projects
+
 # 0.0.1 (2021-10-15)
 
 - **FEATURES:** added grid & cursor